Improve interwiki entries for Wikimedia Foundation projects
The entries have been separated into other block in the SQL script for
readability, this does not matter with respect to the database. Old
entries were removed or replaced with the new site's URL and more
meaningful prefix. There are also couple of convenient aliases for the
most frequently used entries.