Improve instructions for extensions/ArchInterWiki.sql
The instructions for interlanguage links were generalized to include all
interwiki links as a preparation for merging the
maintenance/arch-interwiki.sql script. The API queries were improved to
use the limit=max shortcut, for continuation of queries it is better to
just link to upstream because the previously described method is being
deprecated.