Fix the ArchLinux extension/skin
The Vector skin was rewritten for MW 1.35 and our extension/skin has
become incompatible. The way the Vector skin builds a page has changed,
so changing the 'headelement' does not work anymore - it was never an
officially supported hook.

Unfortunately, it seems we always used a somewhat hackish way to place
the Arch navbar, the Vector skin does not provide a suitable hook for
this. For now I hacked VectorTemplate.php to read the 'headelement'
(which now includes only our changes) again. We need to figure out a
better way to manage this extension in the future...

Other changes are simple stylesheet updates:

- Use updated classes for the style overrides
- Fix top margin of the left panel
- Using a relative path in url() is needed to get a https:// link and
  avoid mixed-content errors